requirements/production.txt
-r base.txt

pyldap==2.4.25.1
django-auth-ldap==1.2.8
django-remote-auth-ldap==0.1.4
psycopg2==2.6.1